+++ src/it/java17/org/codehaus/plexus/archivers/spi/ArchiverProviderIT.java
@@ -25,21 +25,21 @@ import java.util.Arrays;
 import java.util.Comparator;
 import java.util.List;
 
-import org.codehaus.plexus.archiver.ArchiveCreation;
 import org.codehaus.plexus.archiver.ArchivedFileSet;
-import org.codehaus.plexus.archiver.ArchivedFileSetSpec;
 import org.codehaus.plexus.archiver.Archiver;
-import org.codehaus.plexus.archiver.ArchiverConfigurer;
-import org.codehaus.plexus.archiver.DuplicateHandling;
 import org.codehaus.plexus.archiver.FileSet;
-import org.codehaus.plexus.archiver.FileSetSpec;
-import org.codehaus.plexus.archiver.PermissionHandling;
-import org.codehaus.plexus.archiver.UnixPermissions;
 import org.codehaus.plexus.archiver.diags.NoOpArchiver;
-import org.codehaus.plexus.archiver.CaseSensitivities;
-import org.codehaus.plexus.archiver.CaseSensitivity;
-import org.codehaus.plexus.archiver.DefaultExcludes;
-import org.codehaus.plexus.archiver.EmptyDirectoryHandling;
+import org.codehaus.plexus.archivers.config.ArchiveCreation;
+import org.codehaus.plexus.archivers.config.ArchivedFileSetSpec;
+import org.codehaus.plexus.archivers.config.ArchiverConfigurer;
+import org.codehaus.plexus.archivers.config.CaseSensitivities;
+import org.codehaus.plexus.archivers.config.CaseSensitivity;
+import org.codehaus.plexus.archivers.config.DefaultExcludes;
+import org.codehaus.plexus.archivers.config.DuplicateHandling;
+import org.codehaus.plexus.archivers.config.EmptyDirectoryHandling;
+import org.codehaus.plexus.archivers.config.FileSetSpec;
+import org.codehaus.plexus.archivers.config.PermissionHandling;
+import org.codehaus.plexus.archivers.config.UnixPermissions;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
