+++ src/it/java17/org/codehaus/plexus/archivers/spi/PlexusIoResourceCollectionProviderIT.java
@@ -23,16 +23,16 @@ import java.nio.file.Path;
 import java.util.Arrays;
 import java.util.List;
 
-import org.codehaus.plexus.archiver.PlexusIoResourceCollectionConfigurer;
-import org.codehaus.plexus.archiver.SymbolicLinkHandling;
-import org.codehaus.plexus.archiver.gzip.PlexusIoGzipResourceCollectionProvider;
 import org.codehaus.plexus.archiver.Archiver;
-import org.codehaus.plexus.archiver.ArchiverConfigurer;
-import org.codehaus.plexus.archiver.CaseSensitivity;
-import org.codehaus.plexus.archiver.DefaultExcludes;
-import org.codehaus.plexus.archiver.EmptyDirectoryHandling;
+import org.codehaus.plexus.archiver.gzip.PlexusIoGzipResourceCollectionProvider;
 import org.codehaus.plexus.archiver.resources.PlexusIoFileResourceCollectionProvider;
 import org.codehaus.plexus.archiver.zip.PlexusArchiverZipFileResourceCollectionProvider;
+import org.codehaus.plexus.archivers.config.ArchiverConfigurer;
+import org.codehaus.plexus.archivers.config.CaseSensitivity;
+import org.codehaus.plexus.archivers.config.DefaultExcludes;
+import org.codehaus.plexus.archivers.config.EmptyDirectoryHandling;
+import org.codehaus.plexus.archivers.config.PlexusIoResourceCollectionConfigurer;
+import org.codehaus.plexus.archivers.config.SymbolicLinkHandling;
 import org.codehaus.plexus.components.io.resources.AbstractPlexusIoArchiveResourceCollection;
 import org.codehaus.plexus.components.io.resources.AbstractPlexusIoResourceCollection;
 import org.codehaus.plexus.components.io.resources.PlexusIoCompressedFileResourceCollection;
