+++ src/main/java17/org/codehaus/plexus/archiver/manager/PlexusIoResourceCollectionFactory.java
@@ -19,7 +19,7 @@ package org.codehaus.plexus.archiver.manager;
 
 import java.util.function.Consumer;
 
-import org.codehaus.plexus.archiver.PlexusIoResourceCollectionConfigurer;
+import org.codehaus.plexus.archivers.config.PlexusIoResourceCollectionConfigurer;
 import org.codehaus.plexus.components.io.resources.PlexusIoResourceCollection;
 
 /**
